% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absErr.R
\name{absErr}
\alias{absErr}
\title{patient-level prediction error}
\usage{
absErr(datval, model, label.col)
}
\arguments{
\item{datval}{#data on which validation will be computed}

\item{model}{#model!}

\item{label.col}{#column number with labels}
}
\description{
patient-level prediction error
}
