% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_revolving.R
\name{benchmark.revolving}
\alias{benchmark.revolving}
\title{Benchmark for rolling events with history/past/last timeframes vs one-hot encoding for rolling events (patients with several labels at different timestamps)}
\usage{
benchmark.revolving(
  MLHO.dat,
  labels,
  binarize = FALSE,
  sparsity = NA,
  jmi = TRUE,
  topn = 200,
  patients,
  multicore = FALSE,
  valuesToMerge = FALSE,
  timeBufffer = c(h = 0, p = 0, l = 0)
)
}
\arguments{
\item{MLHO.dat}{your dbmart goes here}

\item{labels}{your labeldt goes here}

\item{binarize}{if you want the outcome to be binary}

\item{sparsity}{if you want to apply sparsity}

\item{jmi}{if you want to do jmi}

\item{topn}{the number of features to be selected}

\item{patients}{vector of patients}

\item{multicore}{if you want to parallelize the jmi}

\item{valuesToMerge}{set to true if you have a "value" column for phenx and want to unite these columns into one}

\item{encounterLevel}{set to true if you have multiple date assigned labels per patient}

\item{timeBuffer}{a vector containing three numbers (h,p,l) describing the time buffer that should be added to the dates used to calculate the last, past and history event groups (only used when encounterLevel=TRUE).}
}
\description{
Benchmark for rolling events with history/past/last timeframes vs one-hot encoding for rolling events (patients with several labels at different timestamps)
}
