% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias.R
\name{bias}
\alias{bias}
\title{Bias evaluation metrics}
\usage{
bias(
  group = "all",
  month = "all",
  err,
  model.name,
  keys,
  plot = FALSE,
  calibrate = "RAW"
)
}
\arguments{
\item{group}{#demographic or patient groups, defaults at all}

\item{month}{#temporal segmentation of the data, defaults at all -- i.e., no time stamps}

\item{err}{##the error file generated from mlho learn}

\item{model.name}{##name of the model}

\item{keys}{##phenotype and model information for merging extracted from err}

\item{plot}{##wanna plot out reliability diagrams?}

\item{calibrate}{##wanna calibrate the computed probabilities?}
}
\description{
Bias evaluation metrics
}
