% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliberatio.R
\name{caliber}
\alias{caliber}
\title{Computes calibration data for classification}
\usage{
caliber(datval, model, label.col)
}
\arguments{
\item{datval}{testing data}

\item{model}{model}

\item{label.col}{the column number with labels}
}
\value{
a data frame with calibration information
}
\description{
Computes calibration data for classification
}
\examples{
caliber(datval = dat.VAL,model=BOS.model400,label.col = which( colnames(dat.VAL)=="label" ))
}
