% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datebuff.R
\name{datebuff}
\alias{datebuff}
\title{datebuff}
\usage{
datebuff(uniqpats, dbmart, dems, eve)
}
\arguments{
\item{uniqpats}{a vector of unique patients numbers}

\item{dbmart}{the datamart according to}

\item{dems}{the table including the demographic and event date/status data}

\item{eve}{is the event of interest}
}
\value{
the updated data mart
}
\description{
datebuff
}
\examples{
see vingetes
}
