% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn.R
\name{learn}
\alias{learn}
\title{the ML part}
\usage{
learn(
  data,
  demographics,
  augment = FALSE,
  tst.size = 0.2,
  save.model = FALSE,
  classifier = "glmboost",
  note = "mlho_terst_run",
  aoi,
  multicore = FALSE
)
}
\arguments{
\item{data}{the dbmart}

\item{demographics}{demographics table}

\item{augment}{if you want to add demographic variables to your model set it to TRUE}

\item{tst.size}{test/train set ratio -- default set at 0.2 (20 percent)}

\item{save.model}{do you want to save the model?}

\item{classifier}{the classification algorithm}

\item{note}{add a descriptive word about your modeling}

\item{aoi}{outcome of interest}

\item{multicore}{do you want to parallelize the process?}
}
\description{
the ML part
}
