% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrix.R
\name{metrix}
\alias{metrix}
\title{the MLHO metrix function}
\usage{
metrix(
  datval,
  model,
  label.col,
  note,
  phenx = "phenotype",
  topn,
  class = "binary"
)
}
\arguments{
\item{datval}{data on which validation will be computed}

\item{model}{the model}

\item{label.col}{the column number with labels}

\item{note}{model description}

\item{phenx}{outcome of the study or phenotype of the interest -- default is "phenotype"}

\item{topn}{number of features used in the training data}

\item{class}{binary or continuous prediction}
}
\value{
output includes performance model metrics
}
\description{
the MLHO metrix function
}
