% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlearn.R
\name{mlearn}
\alias{mlearn}
\title{the MLHO ML training and testing}
\usage{
mlearn(
  dat.train,
  dat.test,
  dems = NULL,
  save.model = FALSE,
  classifier = "glmboost",
  note = "mlho_terst_run",
  cv = "cv",
  nfold = 5,
  aoi = "label",
  multicore = FALSE,
  preProc = TRUE,
  calSHAP = FALSE,
  n_incidence = 25,
  counterfactual = FALSE,
  save.model.counterfactual = FALSE
)
}
\arguments{
\item{dat.train}{training data}

\item{dat.test}{testing data}

\item{dems}{demographic variables}

\item{save.model}{do you want to save the model?}

\item{classifier}{the classification algorithm}

\item{note}{add a descriptive word about your modeling}

\item{cv}{cross validatio method}

\item{nfold}{nfold cross validation folds}

\item{aoi}{outcome of interest}

\item{multicore}{do you want to parallelize the process?}

\item{preProc}{preprocessig on the train data or not}

\item{calSHAP}{calculating SHAP value or not, the default is false}

\item{n_incidence}{number of incidences picked for calculating SHAP value}

\item{counterfactual}{models for counterfactual analysis}

\item{save.model.counterfactual}{do you want to save the counterfactual model?}
}
\description{
the MLHO ML training and testing
}
