% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlho.it.R
\name{mlho.it}
\alias{mlho.it}
\title{iterative modeling with MLHO}
\usage{
mlho.it(
  dbmart,
  labels = labeldt,
  dems = NULL,
  test.sample = 30,
  MSMR.binarize = FALSE,
  MSMR.sparsity = 0.005,
  MSMR.jmi = TRUE,
  MSMR.topn = 200,
  mlearn.save.model = FALSE,
  mlearn.note = "mlho_phewas run",
  mlearn.aoi = "demo",
  mlearn.cv = "cv",
  mlearn.nfold = 5,
  multicore = FALSE,
  preProc = TRUE,
  iterations = 5
)
}
\arguments{
\item{dbmart}{dbmart table}

\item{labels}{should be the labeldt table}

\item{dems}{table containing the demographic variables}

\item{test.sample}{put 20 if you want to use 20 percent for testing and 80 percent for training}

\item{MSMR.binarize}{MSMR.lite parameter}

\item{MSMR.sparsity}{MSMR.lite parameter}

\item{MSMR.jmi}{MSMR.lite parameter}

\item{MSMR.topn}{MSMR.lite parameter}

\item{mlearn.save.model}{mlearn parameter}

\item{mlearn.note}{mlearn parameter}

\item{mlearn.aoi}{mlearn parameter}

\item{mlearn.cv}{mlearn parameter}

\item{mlearn.nfold}{mlearn parameter}

\item{multicore}{if you want to parallelize the process}

\item{preProc}{preprocessig on the train data or not}

\item{iterations}{number of iterations you want. recommended at least 5. needs to be numeric}
}
\description{
iterative modeling with MLHO
}
