% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mshapviz_all.R
\name{mshapviz_all}
\alias{mshapviz_all}
\title{SHAP values visualization}
\usage{
mshapviz_all(shap_value, dbmart.concepts, top_n = 10)
}
\arguments{
\item{shap_value}{SHAP values from `mlearn`}

\item{dbmart.concepts}{A table with feature description}

\item{top_n}{The number of top features to display.}
}
\description{
SHAP values visualization
}
\examples{
# Assuming `shap_value` is your data frame with SHAP values
mshapviz_all(shap_value, dbmart.concepts, top_n = 10)
}
