% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMR.lite.R
\name{MSMR.lite}
\alias{MSMR.lite}
\title{MSMR lite}
\usage{
MSMR.lite(
  MLHO.dat = dat.train,
  labels = test_label,
  binarize = FALSE,
  sparsity = NA,
  jmi = TRUE,
  topn = 200,
  patients = uniqpats.train,
  multicore = FALSE,
  encounterLevel = FALSE,
  valuesToMerge = FALSE,
  timeBufffer = c(h = 0, p = 0, l = 0, o = 0)
)
}
\arguments{
\item{MLHO.dat}{your dbmart goes here}

\item{labels}{your labeldt goes here}

\item{binarize}{if you want the outcome to be binary}

\item{sparsity}{if you want to apply sparsity}

\item{jmi}{if you want to do jmi}

\item{topn}{the number of features to be selected}

\item{patients}{vector of patients}

\item{multicore}{if you want to parallelize the jmi}

\item{encounterLevel}{set to true if you have multiple date assigned labels per patient}

\item{valuesToMerge}{set to true if you have a "value" column for phenx and want to unite these columns into one}
}
\description{
MSMR lite
}
