% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMR.lite.R
\name{MSMSR.lite}
\alias{MSMSR.lite}
\title{MSMR lite}
\usage{
MSMSR.lite(
  MLHO.dat,
  labels,
  binarize = FALSE,
  sparsity = NA,
  jmi = TRUE,
  topn = 200,
  patients,
  multicore = TRUE
)
}
\arguments{
\item{MLHO.dat}{## your dbmart}

\item{labels}{##your labeldt}

\item{binarize}{## do you want the output matrix be binarized?}

\item{sparsity}{## step 1 of MSMR. 0.1 = 10% prevalence as minimum}

\item{jmi}{## perform the joint mutual information}

\item{topn}{## what is the top n features you want after jmi?}

\item{patients}{## vector of the unique patients}

\item{multicore}{## do you want to paralelize the compute?}
}
\value{
##AVR training set
}
\description{
MSMR lite
}
