% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasc5.R
\name{pasc5}
\alias{pasc5}
\title{Title}
\usage{
pasc5(
  PatientObservations,
  PatientClinicalCourse,
  PatientSummary,
  loyaltydays1 = 90
)
}
\arguments{
\item{PatientObservations}{2.1 PatientObservations table should include 4-digit ICD codes}

\item{PatientClinicalCourse}{2.1 PatientClinicalCourse table}

\item{PatientSummary}{2.1 PatientSummary table}

\item{loyaltydays1}{works as the minimum days before and after hospitalization to be assumed loyal}
}
\value{
plots for cases 1-4 and data for cases 4 and 5
}
\description{
Title
}
