% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasccases.R
\name{pasccases}
\alias{pasccases}
\title{5 PASC cases}
\usage{
pasccases(
  PatientObservations,
  PatientClinicalCourse,
  PatientSummary,
  loyaltydays1 = 90
)
}
\arguments{
\item{PatientObservations}{2.1 PatientObservations table should include 4-digit ICD codes}

\item{PatientClinicalCourse}{2.1 PatientClinicalCourse table}

\item{PatientSummary}{2.1 PatientSummary table}

\item{loyaltydays1}{work as the minimum days before and after hospitalization to be assumed loyal}
}
\value{
plots for cases 1-4 and data for cases 4 and 5
}
\description{
5 PASC cases
}
