% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mshapviz_all.R
\name{position_bee}
\alias{position_bee}
\title{SHAP values visualization}
\usage{
position_bee(width = NULL, adjust = NULL)
}
\arguments{
\item{data.shap}{Model output containing SHAP values from `mlearn`}

\item{data.descp}{A table with feature description}

\item{top_n}{The number of top features to display.}
}
\description{
This function takes the output from the `mlearn` function and creates either waterfall plot or force plot
showing the top n SHAP values for the specified number of features.
}
\examples{
# Assuming `shap_value` is your data frame with SHAP values
mshapviz_all(shap_value, dbmart.concepts, top_n = 10)
}
