% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_shap.R
\name{visualize_shap}
\alias{visualize_shap}
\title{SHAP values visualization}
\usage{
visualize_shap(
  shap_value,
  dbmart.concepts,
  plot_type = "force",
  top_n = 5,
  num = 1
)
}
\arguments{
\item{plot_type}{'waterfall' or 'force'}

\item{top_n}{The number of top features to display.}

\item{num}{The nth of incidence picked}

\item{data.shap}{Model output containing SHAP values from `mlearn`}

\item{data.descp}{A table with feature description}
}
\description{
This function takes the output from the `mlearn` function and creates either waterfall plot or force plot
showing the top n SHAP values for the specified number of features.
}
\examples{
# Assuming `shap_value` is your data frame with SHAP values
visualize_shap(shap_value, dbmart.concepts, plot_type = "waterfall", top_n = 5, num = 6)
}
