% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wideout.R
\name{wideout}
\alias{wideout}
\title{wideout}
\usage{
wideout(data, binarize = FALSE, sparsity = NA, patients)
}
\arguments{
\item{data}{the dbmart}

\item{binarize}{if you want to use the features as binary, set it to TRUE -- default is FALSE}

\item{sparsity}{if the first step of MSMR, set a threshold between 0 and 1 -- 1 is 100 percent!}

\item{patients}{}
}
\description{
wideout
}
