% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datebuff.R
\name{datebuff}
\alias{datebuff}
\title{datebuff}
\usage{
datebuff(data, demographics, patients, eve, numdays = 0, use)
}
\arguments{
\item{data}{dbmart}

\item{demographics}{demographics table}

\item{patients}{a vector of unique patients numbers}

\item{eve}{the column in the dems table indicating the date for event of interest}

\item{numdays}{number of buffer days -- defaults at 0
#' @param use use data from before or after the buffer days}
}
\value{
the updated data mart
}
\description{
datebuff
}
\examples{
see vignettes
}
