% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSMR.lite.R
\name{MSMSR.lite}
\alias{MSMSR.lite}
\title{MSMR lite}
\usage{
MSMSR.lite(
  MLHO.dat,
  labels,
  binarize = FALSE,
  sparsity = NA,
  jmi = TRUE,
  topn = 200,
  patients,
  multicore = TRUE
)
}
\arguments{
\item{MLHO.dat}{your dbmart goes here}

\item{labels}{your labeldt goes here}

\item{binarize}{if you want the outcome to be binary}

\item{sparsity}{if you want to apply sparsity}

\item{jmi}{if you want to do jmi}

\item{topn}{the number of features to be selected}

\item{patients}{vector of patients}

\item{multicore}{if you want to parallelize the jmi}
}
\value{

}
\description{
MSMR lite
}
